Migrations modified to let foreign keys and seeders work correctly

+++ database/migrations/2023_05_23_000746_create_quotes_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('number');
-            $table->integer('created_by')
+            $table->bigInteger('created_by')
                 ->foreign('created_by')
                 ->references('id')
                 ->on('employees');
-            $table->integer('order_id')
+            $table->bigInteger('order_id')
                 ->nullable()
                 ->foreign('order_id')
                 ->references('id')
